Move createSpreadsheetData method to test-helpers directory
